refactor: move resolveTypeList to utils
Move the `resolveTypeList` function to `utils` so it can be reused in
`Component` and `NgModule` decorator handlers.
